Issue #269 - Added VersionManager and TransactionProcessor.
