+++ lib/core/Observer.ts
@@ -1,34 +1,15 @@
-import AnchorFile, { IAnchorFile } from './AnchorFile';
-import BatchFile from './BatchFile';
 import DownloadManager from './DownloadManager';
 import ErrorCode from '../common/ErrorCode';
-import IResolvedTransaction from './interfaces/IResolvedTransaction';
 import ITransaction from '../common/ITransaction';
+import ITransactionUnderProcessing, { TransactionProcessingStatus } from './interfaces/ITransactionUnderProcessing';
 import OperationProcessor from './OperationProcessor';
-import ProtocolParameters from './ProtocolParameters';
 import timeSpan = require('time-span');
+import TransactionProcessor from './interfaces/TransactionProcessor';
 import TransactionStore from './interfaces/TransactionStore';
 import UnresolvableTransactionStore from './interfaces/UnresolvableTransactionStore';
+import VersionManager from './VersionManager';
 import { Blockchain } from './Blockchain';
-import { FetchResultCode } from '../common/FetchResultCode';
-import { Operation } from './Operation';
 import { SidetreeError } from './Error';
-
-/**
- * The state of a transaction that is being processed.
- */
-enum TransactionProcessingStatus {
-  Pending = 'pending',
-  Processsed = 'processed'
-}
-
-/**
- * Data structure for holding a transaction that is being processed and its state.
- */
-interface ITransactionUnderProcessing {
-  transaction: ITransaction;
-  processingStatus: TransactionProcessingStatus;
-}
 
 /**
  * Class that performs periodic processing of batches of Sidetree operations anchored to the blockchain.
@@ -44,7 +25,7 @@ export default class Observer {
   /**
    * The list of transactions that are being downloaded or processed.
    */
-  private transactionsUnderProcessing: { transaction: ITransaction; processingStatus: TransactionProcessingStatus }[] = [];
+  private transactionsUnderProcessing: ITransactionUnderProcessing[] = [];
 
   /**
    * This is the transaction that is used as a timestamp to fetch newer transaction.
@@ -52,6 +33,7 @@ export default class Observer {
   private lastKnownTransaction: ITransaction | undefined;
 
   public constructor (
+    private versionManager: VersionManager,
     private blockchain: Blockchain,
     private downloadManager: DownloadManager,
     private operationProcessor: OperationProcessor,
@@ -87,7 +69,7 @@ export default class Observer {
 
   /**
    * Processes new transactions if any, then reprocess a set of unresolvable transactions if any,
-   * then scehdules the next round of processing using the following rules unless `stopPeriodicProcessing()` is invoked.
+   * then scehdules the next round of processing unless `stopPeriodicProcessing()` is invoked.
    */
   public async processTransactions () {
     let blockReorganizationDetected = false;
@@ -130,7 +112,7 @@ export default class Observer {
           };
           this.transactionsUnderProcessing.push(awaitingTransaction);
           // Intentionally not awaiting on downloading and processing each operation batch.
-          void this.downloadThenProcessBatchAsync(transaction, awaitingTransaction);
+          void this.processTransaction(transaction, awaitingTransaction);
         }
 
         // If block reorg is detected, we must wait until no more operation processing is pending,
@@ -181,6 +163,7 @@ export default class Observer {
       // Wait a little before checking again.
       await new Promise(resolve => setTimeout(resolve, 1000));
     }
+
     return;
   }
 
@@ -202,7 +185,7 @@ export default class Observer {
       };
       unresolvableTransactionStatus.push(awaitingTransaction);
       // Intentionally not awaiting on downloading and processing each operation batch.
-      void this.downloadThenProcessBatchAsync(transaction, awaitingTransaction);
+      void this.processTransaction(transaction, awaitingTransaction);
     }
 
     // Wait until all unresolvable transactions are processed,
@@ -237,146 +220,33 @@ export default class Observer {
     // Trim the transaction list.
     this.transactionsUnderProcessing.splice(0, i);
   }
+
   /**
-   * Processes the given transaction.
-   * If the given transaction is unresolvable (anchor/batch file not found), save the transaction for retry.
-   * If no error encountered (unresolvable transaction is NOT an error), advance the 'last processed transaction' marker.
+   * Processes the given transaction by passing the transaction to the right version of the transaction processor based on the transaction time.
+   * The transaction processing generically involves first downloading DID operation data from CAS (Content Addressable Storage),
+   * then storing the operations indexed/grouped by DIDs in the persistent operation DB.
    */
-  private async downloadThenProcessBatchAsync (transaction: ITransaction, transactionUnderProcessing: ITransactionUnderProcessing) {
-    let retryNeeded = false;
+  private async processTransaction (transaction: ITransaction, transactionUnderProcessing: ITransactionUnderProcessing) {
+    let transactionProcessedSuccessfully;
 
     try {
-      // Get the protocol parameters
-      const protocolParameters = ProtocolParameters.get(transaction.transactionTime);
-
-      console.info(`Downloading anchor file '${transaction.anchorFileHash}', max size limit ${protocolParameters.maxAnchorFileSizeInBytes} bytes...`);
-      const anchorFileFetchResult = await this.downloadManager.download(transaction.anchorFileHash, protocolParameters.maxAnchorFileSizeInBytes);
-
-      // No thing to process if the file hash is invalid. No retry needed.
-      if (anchorFileFetchResult.code === FetchResultCode.InvalidHash) {
-        console.info(`Anchor file '${transaction.anchorFileHash}' is not a valid hash.`);
-        return;
-      }
-
-      // No thing to process if the file size exceeds protocol specified size limit, no retry needed either.
-      if (anchorFileFetchResult.code === FetchResultCode.MaxSizeExceeded) {
-        console.info(`Anchor file '${transaction.anchorFileHash}' exceeded max size limit ${protocolParameters.maxAnchorFileSizeInBytes} bytes.`);
-        return;
-      }
-
-      // Content for hash exists but is not a file. No retry needed.
-      if (anchorFileFetchResult.code === FetchResultCode.NotAFile) {
-        console.info(`Anchor file hash '${transaction.anchorFileHash}' points to a content that is not a file.`);
-        return;
-      }
-
-      // If Content Addressable Storage is not reachable, mark the transaction for retry later.
-      if (anchorFileFetchResult.code === FetchResultCode.CasNotReachable) {
-        console.info(`CAS not reachable for anchor file '${transaction.anchorFileHash}', will try again later.`);
-        retryNeeded = true;
-        return;
-      }
-
-      // If file cannot be found, mark it for retry later.
-      if (anchorFileFetchResult.code === FetchResultCode.NotFound) {
-        console.info(`Anchor file '${transaction.anchorFileHash}' not found, will try again later.`);
-        retryNeeded = true;
-        return;
-      }
-
-      console.info(`Anchor file '${transaction.anchorFileHash}' of size ${anchorFileFetchResult.content!.length} bytes downloaded.`);
-      let anchorFile: IAnchorFile;
-      try {
-        const maxOperationsPerBatch = protocolParameters.maxOperationsPerBatch;
-        const hashAlgorithmInMultihashCode = protocolParameters.hashAlgorithmInMultihashCode;
-        anchorFile = AnchorFile.parseAndValidate(anchorFileFetchResult.content!, maxOperationsPerBatch, hashAlgorithmInMultihashCode);
-      } catch (error) {
-        // Give meaningful/specific error code and message when possible.
-        if (error instanceof SidetreeError) {
-          console.info(`Invalid anchor file: ${error}`);
-          console.info(`Anchor file '${transaction.anchorFileHash}' failed parsing/validation, transaction '${transaction.transactionNumber}' ignored...`);
-        } else {
-          console.error(`Unexpected error processing anchor file, MUST investigate and fix: ${error}`);
-          retryNeeded = true;
-        }
-
-        return;
-      }
-
-      console.info(`Downloading batch file '${anchorFile.batchFileHash}', max size limit ${protocolParameters.maxBatchFileSizeInBytes}...`);
-      const batchFileFetchResult = await this.downloadManager.download(anchorFile.batchFileHash, protocolParameters.maxBatchFileSizeInBytes);
-
-      // Nothing to process if the file hash is invalid. No retry needed.
-      if (batchFileFetchResult.code === FetchResultCode.InvalidHash) {
-        console.info(`Batch file '${anchorFile.batchFileHash}' is not a valid hash.`);
-        return;
-      }
-
-      // Nothing to process if the file size exceeds protocol specified size limit, no retry needed either.
-      if (batchFileFetchResult.code === FetchResultCode.MaxSizeExceeded) {
-        console.info(`Batch file '${anchorFile.batchFileHash}' exceeded max size limit ${protocolParameters.maxBatchFileSizeInBytes}...`);
-        return;
-      }
-
-      // Content for hash exists but is not a file. No retry needed.
-      if (batchFileFetchResult.code === FetchResultCode.NotAFile) {
-        console.info(`Batch file hash '${anchorFile.batchFileHash}' points to a content that is not a file.`);
-        return;
-      }
-
-      // If Content Addressable Storage is not reachable, mark the transaction for retry later.
-      if (batchFileFetchResult.code === FetchResultCode.CasNotReachable) {
-        console.info(`CAS not reachable for batch file '${anchorFile.batchFileHash}', will try again later.`);
-        retryNeeded = true;
-        return;
-      }
-
-      // If file cannot be found, mark it for retry later.
-      if (batchFileFetchResult.code === FetchResultCode.NotFound) {
-        console.info(`Batch file '${anchorFile.batchFileHash}' not found, will try again later.`);
-        retryNeeded = true;
-        return;
-      }
-
-      console.info(`Batch file '${anchorFile.batchFileHash}' of size ${batchFileFetchResult.content!.length} downloaded.`);
-
-      // Construct a resolved transaction from the original transaction object now that batch file is fetched.
-      const resolvedTransaction: IResolvedTransaction = {
-        transactionNumber: transaction.transactionNumber,
-        transactionTime: transaction.transactionTime,
-        transactionTimeHash: transaction.transactionTimeHash,
-        anchorFileHash: transaction.anchorFileHash,
-        batchFileHash: anchorFile.batchFileHash
-      };
-
-      let operations: Operation[];
-      try {
-        operations = await BatchFile.parseAndValidate(batchFileFetchResult.content!, anchorFile, resolvedTransaction);
-      } catch (error) {
-        console.info(error);
-        console.info(`Batch file '${anchorFile.batchFileHash}' failed parsing/validation, transaction '${transaction.transactionNumber}' ignored.`);
-        return;
-      }
-
-      // If the code reaches here, it means that the batch of operations is valid, process the operations.
-      const endTimer = timeSpan();
-      await this.operationProcessor.process(operations);
-      console.info(`Processed batch '${anchorFile.batchFileHash}' of ${operations.length} operations. Time taken: ${endTimer.rounded()} ms.`);
+      const transactionProcessor: TransactionProcessor = await this.versionManager.getTransactionProcessor(transaction.transactionTime);
+      transactionProcessedSuccessfully = await transactionProcessor.processTransaction(transaction);
     } catch (error) {
       console.error(`Unhandled error encoutnered processing transaction '${transaction.transactionNumber}'.`);
       console.error(error);
-      retryNeeded = true;
+      transactionProcessedSuccessfully = false;
     } finally {
       // Purposely setting processing status first before rest of the code to prevent any possibility of deadlocking the Observer.
       console.info(`Finished processing transaction '${transaction.transactionNumber}'.`);
       transactionUnderProcessing.processingStatus = TransactionProcessingStatus.Processsed;
 
-      if (retryNeeded) {
-        console.info(`Recording failed processing attempt for transaction '${transaction.transactionNumber}'...`);
-        await this.unresolvableTransactionStore.recordUnresolvableTransactionFetchAttempt(transaction);
-      } else {
+      if (transactionProcessedSuccessfully) {
         console.info(`Removing transaction '${transaction.transactionNumber}' from unresolvable transactions if exists...`);
         await this.unresolvableTransactionStore.removeUnresolvableTransaction(transaction);
+      } else {
+        console.info(`Recording failed processing attempt for transaction '${transaction.transactionNumber}'...`);
+        await this.unresolvableTransactionStore.recordUnresolvableTransactionFetchAttempt(transaction);
       }
     }
   }
