+++ tests/core/VersionManager.spec.ts
@@ -1,0 +1,13 @@
+// import VersionManager from '../../lib/core/VersionManager';
+// import { TransactionProcessingStatus } from '../../lib/core/interfaces/ITransactionUnderProcessing';
+
+describe('VersionManager', () => {
+  it('should fetch right Transaction Processor given the transaction time.', async () => {
+    // const versionManager = new VersionManager();
+
+    // const transactionProcessor = await versionManager.getTransactionProcessor(1);
+
+    // const transaction = { anchorFileHash: 'asdf', transactionNumber: 1, transactionTime: 1, transactionTimeHash: 'asdff' };
+    // transactionProcessor.processTransaction(transaction, { processingStatus: TransactionProcessingStatus.Pending, transaction });
+  });
+});
